api/requirements.txt
absl-py==0.11.0
alembic==1.5.2
astunparse==1.6.3
boto3==1.16.59
botocore==1.19.59
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.1
pipfile==0.0.2
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
